Change mouse coordinates type from double to long.
2008-12-02  Behdad Esfahbod  <behdad@gnome.org>

        * src/vte-private.h:
        * src/vte.c (vte_terminal_send_mouse_button_internal),
        (vte_terminal_maybe_send_mouse_drag), (cursor_inside_match),
        (vte_terminal_match_hilite_show),
        (vte_terminal_match_hilite_update), (vte_terminal_match_hilite),
        (vte_terminal_extend_selection), (vte_terminal_autoscroll),
        (vte_terminal_motion_notify), (vte_terminal_button_press),
        (vte_terminal_button_release):
        Change mouse coordinates type from double to long.


svn path=/trunk/; revision=2299

+++ src/vte-private.h
@@ -292,7 +292,7 @@ struct _VteTerminalPrivate {
 		selection_type_line
 	} selection_type;
 	struct selection_event_coords {
-		double x, y;
+		long x, y;
 	} selection_origin, selection_last;
 	struct selection_cell_coords {
 		long row, col;
@@ -342,7 +342,7 @@ struct _VteTerminalPrivate {
 #define MOUSE_EVENT_ALL_MOTION_TRACKING		0x0010
 	gint mouse_event_mode;
 	guint mouse_last_button;
-	gdouble mouse_last_x, mouse_last_y;
+	long mouse_last_x, mouse_last_y;
 	gboolean mouse_autohide;
 	guint mouse_autoscroll_tag;
